spelling: Add stripped whitespace
This is a classical two-column hexdump.
The last character in the text is a space (20) and it needs to be
present in the presented content as it is a real character.

This distinguishes it from the last line in a block which may be
fractional and not extend the full 16-column width.

Signed-off-by: Josh Soref <2119212+jsoref@users.noreply.github.com>